Wrote test for MatchProvider.
