+++ modules/__tests__/Match-test.js
@@ -1,5 +1,6 @@
 import expect from 'expect'
 import React, { PropTypes } from 'react'
+import MatchProvider from '../MatchProvider'
 import Match from '../Match'
 import { renderToString } from 'react-dom/server'
 import { render } from 'react-dom'
@@ -284,23 +285,6 @@ describe('Match', () => {
       const html = renderToString(
         <LocationProvider location={location}>
           <Match pattern="/" render={({ location }) => (
-            <div>{location.state.test}</div>
-          )}/>
-        </LocationProvider>
-      )
-      expect(html).toContain(TEXT)
-    })
-
-
-    it('works with multiple routes', () => {
-      const TEXT = 'TEXT'
-      const location = { pathname: '/', state: { test: TEXT } }
-      const html = renderToString(
-        <LocationProvider location={location}>
-          <Match pattern="/" render={({ location }) => (
-            <div>{location.state.test}</div>
-          )}/>
-          <Match exactly pattern="/myTest" render={({ location }) => (
             <div>{location.state.test}</div>
           )}/>
         </LocationProvider>
